+++ Samples/src/StateVisualizer/VisualDebugger.csproj
